+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/StringTableValue.java
@@ -1,5 +1,6 @@
 package com.kaylerrenslow.armaDialogCreator.arma.stringtable;
 
+import javafx.collections.ObservableMap;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
@@ -12,8 +13,9 @@ public interface StringTableValue {
 	@NotNull
 	StringTableKey getKey();
 
+
 	@NotNull
-	Map<Language, String> getValues();
+	ObservableMap<Language, String> getLanguageTokenMap();
 
 	default boolean equalsValue(StringTableValue other) {
 		if (other == null) {
@@ -22,9 +24,9 @@ public interface StringTableValue {
 		if (other == this) {
 			return true;
 		}
-		for (Map.Entry<Language, String> myEntry : getValues().entrySet()) {
+		for (Map.Entry<Language, String> myEntry : getLanguageTokenMap().entrySet()) {
 			boolean match = false;
-			for (Map.Entry<Language, String> otherEntry : other.getValues().entrySet()) {
+			for (Map.Entry<Language, String> otherEntry : other.getLanguageTokenMap().entrySet()) {
 				if (myEntry.getKey().equals(otherEntry.getKey())) {
 					if (myEntry.getValue().equals(otherEntry.getValue())) {
 						match = true;