+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/impl/StringTableValueImpl.java
@@ -3,18 +3,17 @@ package com.kaylerrenslow.armaDialogCreator.arma.stringtable.impl;
 import com.kaylerrenslow.armaDialogCreator.arma.stringtable.Language;
 import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableKey;
 import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableValue;
+import javafx.collections.ObservableMap;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.Map;
 
 /**
  @author Kayler
  @since 12/12/2016 */
 public class StringTableValueImpl implements StringTableValue {
 	private final StringTableKey key;
-	private final Map<Language, String> values;
+	private final ObservableMap<Language, String> values;
 
-	public StringTableValueImpl(@NotNull StringTableKey key, @NotNull Map<Language, String> values) {
+	public StringTableValueImpl(@NotNull StringTableKey key, @NotNull ObservableMap<Language, String> values) {
 		this.key = key;
 		this.values = values;
 	}
@@ -25,9 +24,9 @@ public class StringTableValueImpl implements StringTableValue {
 		return key;
 	}
 
-	@Override
 	@NotNull
-	public Map<Language, String> getValues() {
+	@Override
+	public ObservableMap<Language, String> getLanguageTokenMap() {
 		return values;
 	}