+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/actions/mainMenu/TestAction.java
@@ -1,9 +1,12 @@
 package com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu;
 
-import com.kaylerrenslow.armaDialogCreator.gui.fx.notification.Notification;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.notification.Notifications;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTable;
+import com.kaylerrenslow.armaDialogCreator.data.io.xml.DefaultStringTableXmlParser;
+import com.kaylerrenslow.armaDialogCreator.gui.fx.main.stringtable.StringTableEditorPopup;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
+
+import java.io.File;
 
 /**
  @author Kayler
@@ -14,15 +17,12 @@ public class TestAction implements EventHandler<ActionEvent> {
 
 	@Override
 	public void handle(ActionEvent event) {
-		if (i % 2 == 0) {
-			Notifications.showNotification(new Notification("hello" + (i++),
-							"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce vestibulum nunc id ligula suscipit, nec rutrum metus lacinia. Nulla vestibulum, orci vitae dapibus elementum, leo.",
-							6 * 1000
-					)
-			);
-		} else {
-			Notifications.showNotification(new Notification("hello" + (i++), "this is a test <a href=\"google.com\">anchor</a>", 6 * 1000));
-
+		try {
+			StringTable table = new DefaultStringTableXmlParser(new File("D:\\Archive\\Intellij Files\\Arma Tools\\Arma Dialog " +
+					"Creator\\tests\\com\\kaylerrenslow\\armaDialogCreator\\data\\io\\xml\\stringtable.xml")).createStringTableInstance();
+			new StringTableEditorPopup(table).show();
+		} catch (Exception e) {
+			e.printStackTrace();
 		}
 	}
 }